src/vaultlint/checks/structure/struct_checker.py
"""Example basic vault structure check."""

import logging
from pathlib import Path
from typing import TYPE_CHECKING

import yaml

if TYPE_CHECKING:  # pragma: no cover
    from vaultlint.cli import LintContext

LOG = logging.getLogger("vaultlint.checks.structure_checker")  # pragma: no cover


def load_spec_file(path: str):
    """Load a YAML file and return its content as a dictionary.

    Args:
        path: Path to the YAML specification file

    Returns:
        dict: Parsed YAML content

    Raises:
        FileNotFoundError: If the file doesn't exist
        PermissionError: If there are permission issues accessing the file
        yaml.YAMLError: If the YAML content is invalid
        OSError/IOError: For other I/O related errors
    """
    path = Path(path)

    try:
        if not path.exists():
            raise FileNotFoundError(f"File not found: {path}")

        with path.open("r", encoding="utf-8") as f:
            data = yaml.safe_load(f)

        return data

    except yaml.YAMLError as e:
        LOG.error("YAML error while processing specification file: %s", e)
        raise
    except FileNotFoundError as e:
        LOG.error("Specification file not found: %s", e)
        raise
    except PermissionError as e:
        LOG.error("Permission denied when accessing specification file: %s", e)
        raise
    except (OSError, IOError) as e:
        LOG.error("I/O error while reading specification file: %s", e)
        raise


def struct_checker(context: "LintContext") -> bool:
    """Check vault structure against specification.

    Args:
        context: LintContext containing vault_path and spec_path

    Returns:
        bool: True if structure check passed, False otherwise
    """
    if context.spec_path is None:
        # No spec file - this is normal, not an error
        return True  # Consider this a success if no spec is required

    try:
        spec = load_spec_file(str(context.spec_path))
        # Keep only debug logging for development
        LOG.debug("Loaded specification from: %s", context.spec_path)
        LOG.debug("Spec content: %s", spec)

        # TODO: Implement actual structure validation logic here
        # For now, just validate that we can load the spec

        return True  # Placeholder - return True if spec loads successfully

    except (yaml.YAMLError, FileNotFoundError, PermissionError, OSError, IOError):
        # All specific errors are already logged by load_spec_file()
        # Just return False to indicate validation failure
        return False
